Level Layout v1.1
Level layout is most fleshed out with enemies sprinkled around. Once enemy prefabs and scripts are created we can replace them.

Puzzle 2 has been conceptualized and partially completed. Will need more work on it but I am pushing this so that other people can work and not be too stressed about it.