Lucas 3AM SleepyPatch
Updated all sprites to have the proper settings that were posted in discord

Added proper sprite to the button prefab

Added in boss in a very rough way. Ethan will have to check if it is working as intended

Added scripts for loading new scenes, increasing damage, and increasing health. Small changes to other scripts to supplement this

Enemy layout adjusted